lxc: Use atomic ops for driver->nactive

+++ src/lxc/lxc_conf.h
@@ -75,7 +75,7 @@ struct _virLXCDriver {
 
     virSysinfoDefPtr hostsysinfo;
 
-    size_t nactive;
+    unsigned int nactive;
 
     virStateInhibitCallback inhibitCallback;
     void *inhibitOpaque;
